Refactor move ModalDialog tests
